Add whitelabel db config support
This replaces the previous commit (71ca38bcc), which would not support
caching the config files. This uses the host domain to fetch a list of
config entries from the db, which then replaces the set config at
runtime.
The host is read in middleware, which for now is registered at the top
level, for all registered routes. We may wish to fine-tune that at some
point (so admin pages aren't affected, for example).