:wrench: Move to a map format for feature limits
This allows us to more easily define exactly which routes/methods are
available for certain levels, in one place, which should then also allow
us to dynamically show/hide available features, on demand. Current
caching may be a bit harsh, but should match with the hard-coded nature
of the list.

At a later stage, we'll probably move one or more of the lists to a
config file, or similar.